+++ packages/field/package.json
@@ -2,8 +2,6 @@
   "name": "@ant-design/pro-field",
   "version": "0.0.11",
   "description": "原子信息组件，统一 ProForm、ProTable、ProList、Filter 等组件里面的字段定义。",
-  "main": "lib/index.js",
-  "types": "lib/index.d.ts",
   "homepage": "https://github.com/ant-design/pro-components/tree/master/packages/plugin-layout#readme",
   "bugs": "https://github.com/ant-design/pro-components/issues",
   "repository": {
@@ -11,26 +9,28 @@
     "url": "https://github.com/ant-design/pro-components"
   },
   "license": "MIT",
+  "author": "chenshuai2144",
+  "main": "lib/index.js",
+  "types": "lib/index.d.ts",
   "files": [
     "lib",
     "src",
     "es",
     "dist"
   ],
-  "author": "chenshuai2144",
   "dependencies": {
     "classnames": "^2.2.6",
     "lodash.isequal": "^4.5.0",
     "lodash.tonumber": "^4.0.3",
     "moment": "^2.27.0"
   },
-  "peerDependencies": {
-    "antd": "4.x",
-    "react": "^16.x"
-  },
   "devDependencies": {
     "@types/classnames": "^2.2.10",
     "@types/lodash.isequal": "^4.5.5",
     "@types/lodash.tonumber": "^4.0.6"
+  },
+  "peerDependencies": {
+    "antd": "4.x",
+    "react": "^16.x"
   }
 }
