+++ packages/layout/package.json
@@ -26,12 +26,11 @@
   "dependencies": {
     "@ant-design/icons": "^4.0.0",
     "@ant-design/pro-provider": "0.0.2",
+    "@ant-design/pro-utils":"0.0.1",
     "@umijs/route-utils": "^1.0.5",
     "antd": "^4.2.0",
     "classnames": "^2.2.6",
-    "hash.js": "^1.1.7",
     "history": "^4.10.1",
-    "lodash.isequal": "^4.5.0",
     "memoize-one": "^5.1.1",
     "omit.js": "^2.0.2",
     "path-to-regexp": "2.4.0",
