+++ packages/layout/src/SiderMenu/BaseMenu.tsx
@@ -4,12 +4,13 @@ import { Menu } from 'antd';
 import React, { useEffect, useState, useRef, useMemo } from 'react';
 import classNames from 'classnames';
 import useMergedState from 'rc-util/lib/hooks/useMergedState';
+import { isUrl, isImg } from '@ant-design/pro-utils';
 
 import { MenuMode, MenuProps } from 'antd/es/menu';
 import { MenuTheme } from 'antd/es/menu/MenuContext';
 import defaultSettings, { PureSettings } from '../defaultSettings';
 import { getSelectedMenuKeys } from './SiderMenuUtils';
-import { isUrl, getOpenKeysFromMenuData, isImg } from '../utils/utils';
+import { getOpenKeysFromMenuData } from '../utils/utils';
 
 import { MenuDataItem, MessageDescriptor, Route, RouterTypes, WithFalse } from '../typings';
 import MenuCounter from './Counter';
