+++ packages/layout/src/utils/getBreadcrumbProps.tsx
@@ -2,10 +2,11 @@ import H from 'history';
 import { BreadcrumbProps as AntdBreadcrumbProps } from 'antd/es/breadcrumb';
 import React from 'react';
 import pathToRegexp from 'path-to-regexp';
+import { isBrowser } from '@ant-design/pro-utils';
+
 import { ProSettings } from '../defaultSettings';
 import { MenuDataItem, MessageDescriptor } from '../typings';
 import { urlToList } from './pathTools';
-import { isBrowser } from './utils';
 
 export interface BreadcrumbProps {
   breadcrumbList?: { title: string; href: string }[];
