+++ packages/list/package.json
@@ -28,14 +28,12 @@
     "classnames": "^2.2.6",
     "dnd-core": "^10.0.2",
     "lodash.isequal": "^4.5.0",
-    "lodash.tonumber": "^4.0.3",
     "moment": "^2.24.0",
     "rc-resize-observer": "^0.1.3",
     "rc-util": "^4.19.0",
     "react-dnd": "^10.0.2",
     "react-dnd-html5-backend": "^10.0.2",
     "unstated-next": "^1.1.0",
-    "use-json-comparison": "^1.0.5",
     "use-media-antd-query": "1.0.2"
   },
   "peerDependencies": {
