+++ packages/table/package.json
@@ -27,11 +27,10 @@
     "@ant-design/icons": "^4.1.0",
     "@ant-design/pro-field": "^0.0.9",
     "@ant-design/pro-provider": "0.0.2",
+    "@ant-design/pro-utils":"0.0.1",
     "antd": "^4.1.5",
     "classnames": "^2.2.6",
     "dnd-core": "^10.0.2",
-    "lodash.isequal": "^4.5.0",
-    "lodash.tonumber": "^4.0.3",
     "moment": "^2.24.0",
     "rc-resize-observer": "^0.2.1",
     "rc-util": "^5.0.1",
