--- packages/table/src/component/status/index.less
@@ -1,7 +1,0 @@
-@import '~antd/es/style/themes/default.less';
-
-@pro-dropdown-prefix-cls: ~'@{ant-prefix}-pro-table-dropdown';
-
-.@{pro-dropdown-prefix-cls} {
-  width: auto;
-}
