+++ packages/utils/src/components/LabelIconTip/index.less
@@ -1,8 +1,8 @@
 @import '~antd/es/style/themes/default.less';
 
-@pro-form-label-tip: ~'@{ant-prefix}-pro-form-label-tip';
+@pro-core-label-tip: ~'@{ant-prefix}-pro-core-label-tip';
 
-.@{pro-form-label-tip} {
+.@{pro-core-label-tip} {
   &-icon {
     color: rgba(0, 0, 0, 0.45);
   }
