+++ packages/utils/src/hooks/useDeepCompareEffect/index.ts
@@ -1,0 +1,23 @@
+import React, { useEffect, useRef } from 'react';
+import isEqual from 'lodash.isequal';
+
+function deepCompareEquals(a: any, b: any) {
+  return isEqual(a, b);
+}
+
+function useDeepCompareMemoize(value: any) {
+  const ref = useRef();
+  // it can be done by using useMemo as well
+  // but useRef is rather cleaner and easier
+  if (!deepCompareEquals(value, ref.current)) {
+    ref.current = value;
+  }
+
+  return ref.current;
+}
+
+function useDeepCompareEffect(effect: React.EffectCallback, dependencies?: Object) {
+  useEffect(effect, useDeepCompareMemoize(dependencies));
+}
+
+export default useDeepCompareEffect;
