+++ packages/utils/src/hooks/usePrevious/index.ts
@@ -1,0 +1,13 @@
+import { useEffect, useRef } from 'react';
+
+const usePrevious = <T>(state: T): T | undefined => {
+  const ref = useRef<T>();
+
+  useEffect(() => {
+    ref.current = state;
+  });
+
+  return ref.current;
+};
+
+export default usePrevious;
