+++ packages/utils/src/isBrowser/index.ts
@@ -1,0 +1,7 @@
+const isNode =
+  typeof process !== 'undefined' && process.versions != null && process.versions.node != null;
+
+const isBrowser = () =>
+  typeof window !== 'undefined' && typeof window.document !== 'undefined' && !isNode;
+
+export default isBrowser;
